test: split launcher lifecycle coverage
